requirements.txt
# Core data processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# API and web requests
requests==2.31.0
urllib3==1.26.16

# AWS integration
boto3==1.28.57
botocore==1.31.57

# Google Cloud integration
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0

# Apache Airflow (for local development)
apache-airflow==2.7.1
apache-airflow-providers-google==10.3.0
apache-airflow-providers-amazon==8.5.1

# Data validation
jsonschema==4.19.0
pydantic==2.3.0

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
freezegun==1.2.2

# Development tools
black==23.7.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.3.3

# Utilities
click==8.1.7
tqdm==4.66.1
tabulate==0.9.0